Update ovs-setup.sh
I changed this script to match what I typed in my Debian 12 system. I have a second PC to test this script on, I'll do it later. I noticed that it takes a long time for a computer to get an IP address, so please note this is a project and I am learning as I go. I want network switches that support VLANs and advanced features. If these scripts help someone, then that would be cool. This script is intended to be for myself.